requirements.txt
flask==3.0.2
python-dotenv==1.0.1
openai==1.3.0
apscheduler==3.10.4
requests==2.31.0
python-dateutil==2.8.2
httpx>=0.24.1 